+++ lib/mix/tasks/yacto.migrate.ex
@@ -41,14 +41,6 @@ defmodule Mix.Tasks.Yacto.Migrate do
           end
         end
 
-        # メモ:
-        # マイグレーションディレクトリに存在するディレクトリ（モジュール名）をすべて列挙する
-        # ※各ファイル名にはバージョンとDB名と操作種別が付いている
-        # 各ディレクトリのマイグレーションファイルに対して、
-        #   - ファイル名から DB 名を取り出し、その DB 名が利用する Repo 一覧を調べる
-        #   - その各 Repo に既に入っているバージョンを確認し、そのリポジトリに対してマイグレーション対象のファイルかどうかを調べる
-        # マイグレーション対象のファイルをロードしてマイグレートし、バージョン情報を書き込む
-
         databases = Application.fetch_env!(:yacto, :databases)
 
         {schema_names, messages} = Yacto.Migration.File.list_migration_modules(migration_dir)
@@ -62,7 +54,7 @@ defmodule Mix.Tasks.Yacto.Migrate do
             for message <- messages do
               Logger.warn(message)
             end
-            Yacto.Migration.Migrator.up2(app, repo, String.to_atom(schema_name), migration_dir, migration_files,
+            Yacto.Migration.Migrator.up(app, repo, String.to_atom(schema_name), migration_dir, migration_files,
               db_opts: [databases: databases]
             )
           end
